+++ static/js/apps/eval_rig/eval_section.tsx
@@ -14,31 +14,107 @@
  * limitations under the License.
  */
 
-import React, { useContext, useRef } from "react";
+import React, { useContext, useEffect, useState } from "react";
 import { Button } from "reactstrap";
 
+import { loadSpinner, removeSpinner } from "../../shared/util";
+import {
+  DC_CALL_SHEET,
+  DC_QUESTION_COL,
+  DC_RESPONSE_COL,
+  DC_STAT_COL,
+  LLM_STAT_COL,
+} from "./constants";
 import { AppContext, SessionContext } from "./context";
-import { FeedbackForm, FormStatus } from "./feedback_form";
+import { getCallData, saveToSheet, saveToStore } from "./data_store";
+import { OneQuestion } from "./one_question";
+import { EvalInfo, Response } from "./types";
+
+const LOADING_CONTAINER_ID = "form-container";
+
+export enum FormStatus {
+  NotStarted = 1,
+  InProgress = 2,
+  Completed = 3,
+  Submitted = 4,
+}
+
+const emptyResponse = {
+  overall: "",
+  question: "",
+  llmStat: "",
+  dcResponse: "",
+  dcStat: "",
+};
 
 export function EvalSection(): JSX.Element {
-  const { allQuery, allCall } = useContext(AppContext);
+  const { allQuery, allCall, doc, sheetId, userEmail } = useContext(AppContext);
   const { sessionQueryId, setSessionQueryId, sessionCallId, setSessionCallId } =
     useContext(SessionContext);
 
-  const formRef = useRef(null);
+  const [evalInfo, setEvalInfo] = useState<EvalInfo | null>(null);
+  const [response, setResponse] = useState<Response>(emptyResponse);
+  const [status, setStatus] = useState<FormStatus>(null);
 
-  const submitForm = async (): Promise<boolean> => {
-    return formRef.current.submitForm();
-  };
+  useEffect(() => {
+    getCallData(sheetId, sessionQueryId, sessionCallId).then((data) => {
+      if (data) {
+        setResponse(data as Response);
+        setStatus(FormStatus.Submitted);
+      } else {
+        setResponse(emptyResponse);
+        setStatus(FormStatus.NotStarted);
+      }
+    });
+  }, [sheetId, sessionQueryId, sessionCallId]);
+
+  useEffect(() => {
+    const sheet = doc.sheetsByTitle[DC_CALL_SHEET];
+    if (!(sessionQueryId in allCall)) {
+      setEvalInfo(null);
+      return;
+    }
+    const rowIdx = allCall[sessionQueryId][sessionCallId];
+    sheet.getRows({ offset: rowIdx - 1, limit: 1 }).then((rows) => {
+      const row = rows[0];
+      if (row) {
+        setEvalInfo({
+          dcResponse: row.get(DC_RESPONSE_COL),
+          dcStat: row.get(DC_STAT_COL),
+          llmStat: row.get(LLM_STAT_COL),
+          question: row.get(DC_QUESTION_COL),
+        });
+      }
+    });
+  }, [doc, allCall, sessionQueryId, sessionCallId, setSessionCallId]);
 
   const checkAndSubmit = async (): Promise<boolean> => {
-    const formStatus = formRef.current.status();
-    if (formStatus === FormStatus.InProgress) {
+    if (status === FormStatus.InProgress) {
       alert("Please fill in all fields");
       return false;
     }
-    if (formStatus === FormStatus.Completed) {
-      return submitForm();
+    if (status === FormStatus.Completed) {
+      loadSpinner(LOADING_CONTAINER_ID);
+      return Promise.all([
+        saveToStore(
+          userEmail,
+          sheetId,
+          sessionQueryId,
+          sessionCallId,
+          response
+        ),
+        saveToSheet(userEmail, doc, sessionQueryId, sessionCallId, response),
+      ])
+        .then(() => {
+          return true;
+        })
+        .catch((error) => {
+          alert("Error submitting response: " + error);
+          return false;
+        })
+        .finally(() => {
+          removeSpinner(LOADING_CONTAINER_ID);
+        });
     }
     // Otherwise form status is Submitted or NotStarted. Just proceed with
     // any action.
@@ -48,6 +124,25 @@ export function EvalSection(): JSX.Element {
   const numCalls = () => {
     // Not all queries have calls.
     return Object.keys(allCall[sessionQueryId] || {}).length;
+  };
+
+  const handleChange = (event) => {
+    const { name, value } = event.target;
+    setResponse((prevState) => {
+      const newState = {
+        ...prevState,
+        [name]: value,
+      };
+      let tmpStatus = FormStatus.Completed;
+      for (const value of Object.values(newState)) {
+        if (value === "") {
+          tmpStatus = FormStatus.InProgress;
+          break;
+        }
+      }
+      setStatus(tmpStatus);
+      return newState;
+    });
   };
 
   // Button Actions
@@ -105,52 +200,146 @@ export function EvalSection(): JSX.Element {
     );
   };
 
+  let dcResponseOptions;
+  if (evalInfo) {
+    if (evalInfo.dcStat) {
+      dcResponseOptions = {
+        DC_ANSWER_IRRELEVANT: "Doesn't match the question",
+        DC_ANSWER_RELEVANT: "Relevant and direct",
+      };
+    } else {
+      dcResponseOptions = {
+        DC_ANSWER_EMPTY_BADNL: "Data exists, but NL fails to respond",
+        DC_ANSWER_EMPTY_NODATA: "Query asks for data that doesn't exist in DC",
+        DC_ANSWER_EMPTY_OUTOFSCOPE:
+          "Query asks for data that is out-of-scope for DC",
+      };
+    }
+  }
+
   return (
     <>
-      <FeedbackForm ref={formRef} />
-      {formRef.current && (
-        <div>
-          <span>
-            {sessionCallId} / {numCalls()} ITEMS IN THIS QUERY
-          </span>
-          {showPrevQuery() && (
-            <Button
-              onClick={() => {
-                prevQuery();
-              }}
-            >
-              Previous Query
-            </Button>
-          )}
-          {showPrev() && (
-            <Button
-              onClick={() => {
-                prev();
-              }}
-            >
-              Previous
-            </Button>
-          )}
-          {showNext() && (
-            <Button
-              onClick={() => {
-                next();
-              }}
-            >
-              Next
-            </Button>
-          )}
-          {showNextQuery() && (
-            <Button
-              onClick={() => {
-                nextQuery();
-              }}
-            >
-              Continue to next query
-            </Button>
-          )}
-        </div>
-      )}
+      <div id="form-container">
+        {evalInfo && (
+          <form>
+            <fieldset>
+              <div>
+                <h2>OVERALL EVALUATION</h2>
+                <OneQuestion
+                  question="How is the overall answer?"
+                  name="overall"
+                  options={{
+                    LLM_ANSWER_HALLUCINATION: "Found factual inaccuracies",
+                    LLM_ANSWER_OKAY: "No obvious factual inaccuracies",
+                  }}
+                  handleChange={handleChange}
+                  responseField={response.overall}
+                  disabled={status === FormStatus.Submitted}
+                />
+              </div>
+
+              <div>
+                <h2>GEMMA MODEL EVALUATION</h2>
+                <h3>{evalInfo.question}</h3>
+                <h3>{evalInfo.llmStat}</h3>
+                <OneQuestion
+                  question="Question from the model"
+                  name="question"
+                  options={{
+                    DC_QUESTION_IRRELEVANT:
+                      "Irrelevant, vague, requires editing",
+                    DC_QUESTION_RELEVANT: "Well formulated & relevant",
+                  }}
+                  handleChange={handleChange}
+                  responseField={response.question}
+                  disabled={status === FormStatus.Submitted}
+                />
+                <OneQuestion
+                  question="Model response quality"
+                  name="llmStat"
+                  options={{
+                    LLM_STAT_ACCURATE: "Stats seem accurate",
+                    LLM_STAT_INACCURATE: "Stats seem inaccurate",
+                  }}
+                  handleChange={handleChange}
+                  responseField={response.llmStat}
+                  disabled={status === FormStatus.Submitted}
+                />
+              </div>
+
+              <div>
+                <h2>DATA COMMONS EVALUATION</h2>
+                <h3>{evalInfo.dcResponse}</h3>
+                <h3>{evalInfo.dcStat}</h3>
+                <OneQuestion
+                  question="Response from Data Commons"
+                  name="dcResponse"
+                  options={dcResponseOptions}
+                  handleChange={handleChange}
+                  responseField={response.dcResponse}
+                  disabled={status === FormStatus.Submitted}
+                />
+
+                <OneQuestion
+                  question="Response from Data Commons"
+                  name="dcStat"
+                  options={{
+                    DC_STAT_ACCURATE: "Stats seem accurate",
+                    DC_STAT_INACCURATE: "Stats seem inaccurate",
+                  }}
+                  handleChange={handleChange}
+                  responseField={response.dcStat}
+                  disabled={status === FormStatus.Submitted}
+                />
+              </div>
+            </fieldset>
+          </form>
+        )}
+      </div>
+      <div>
+        <span>
+          {sessionCallId} / {numCalls()} ITEMS IN THIS QUERY
+        </span>
+        {showPrevQuery() && (
+          <Button
+            onClick={() => {
+              prevQuery();
+            }}
+          >
+            Previous Query
+          </Button>
+        )}
+        {showPrev() && (
+          <Button
+            onClick={() => {
+              prev();
+            }}
+          >
+            Previous
+          </Button>
+        )}
+        {showNext() && (
+          <Button
+            onClick={() => {
+              next();
+            }}
+          >
+            Next
+          </Button>
+        )}
+        {showNextQuery() && (
+          <Button
+            onClick={() => {
+              nextQuery();
+            }}
+          >
+            Continue to next query
+          </Button>
+        )}
+      </div>
+      <div id="page-screen" className="screen">
+        <div id="spinner"></div>
+      </div>
     </>
   );
 }
